requirements.txt
# RAG核心依赖
sentence-transformers==2.2.2
chromadb==0.4.15
jieba==0.42.1
pandas==2.1.3
numpy==1.24.3

# EPUB处理
ebooklib==0.18
beautifulsoup4==4.12.2

# API客户端
openai==1.3.0
python-dotenv==1.0.0

# 可选依赖（用于数据分析和可视化）
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0